refactor LadderResetTeamsDialog.cs
